:hammer: Refactor: 반려동물 등록 폼을 feature 단위로 분리
- 반려동물 등록 페이지를 pages 엔트리와 feature 로직으로 분리
- 등록 폼 상태, 검증, 이미지 업로드, 제출 로직을 usePetRegistrationForm 훅으로 이동
- 등록 폼 UI와 이미지 선택 UI를 별도 컴포넌트로 분리
- 만 나이 계산, 상수, 검증 로직을 lib 계층으로 정리

`#29`
