+++ src/features/pet-registration/index.ts
@@ -1,0 +1,2 @@
+export { PetRegistrationForm } from './ui/PetRegistrationForm';
+export { usePetRegistrationForm } from './model/usePetRegistrationForm';
