+++ src/features/pet-registration/lib/helpers.ts
@@ -1,0 +1,19 @@
+export function calculateInternationalAge(birth: string): number | null {
+  if (!birth) return null;
+
+  const today = new Date();
+  const birthDate = new Date(birth);
+
+  if (Number.isNaN(birthDate.getTime())) {
+    return null;
+  }
+
+  let age = today.getFullYear() - birthDate.getFullYear();
+  const monthDiff = today.getMonth() - birthDate.getMonth();
+
+  if (monthDiff < 0 || (monthDiff === 0 && today.getDate() < birthDate.getDate())) {
+    age -= 1;
+  }
+
+  return Math.max(age, 0);
+}
